% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_class_per_layer.R
\name{add_class_per_layer}
\alias{add_class_per_layer}
\title{Add Class Per Layer}
\usage{
add_class_per_layer(layer_predictions, layer)
}
\arguments{
\item{layer_predictions}{The matrix of predictions from classify_cells}

\item{layer}{layer of the model can be layer_1, layer_2_blood, layer_2_cancer, layer_3_TNK, layer_3_MDC, layer_3_BCell, layer_4_CD4_CD8, layer_4_CD8_NK, layer_3_ov_endo, layer_3_breast_lung, layer_3_brain_lung_nbm_skin, layer_4_brain_nbm_skin, layer_5_brain_nbm}
}
\value{
matrix with classes
}
\description{
Add Class Per Layer
}
