% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_unclassified_automatic_threshold.R
\name{add_unclassified_automatic_threshold}
\alias{add_unclassified_automatic_threshold}
\title{Add Unclassified}
\usage{
add_unclassified_automatic_threshold(
  cell_name,
  predictions,
  layer,
  threshold_use
)
}
\arguments{
\item{cell_name}{the name of the cell to score}

\item{predictions}{path to the matrix generated from classify cells and merged with score class}

\item{layer}{layer of the model can be layer_1, layer_2_blood, layer_2_cancer, layer_3_TNK, layer_3_MDC, layer_3_BCell, layer_4_CD4_CD8, layer_4_CD8_NK, layer_3_ov_endo, layer_3_breast_lung, layer_3_brain_lung_nbm_skin, layer_4_brain_nbm_skin, layer_5_brain_nbm}

\item{cutoff}{value for cutoff}
}
\value{
vecotr of filtered classes
}
\description{
Add Unclassified
}
