% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_threshold.R
\name{automatic_threshold}
\alias{automatic_threshold}
\title{Automate Threshold}
\usage{
automatic_threshold(score, unimodal_nsd, bimodal_nsd)
}
\arguments{
\item{score}{Vector containing scores of a class}

\item{unimodal_nsd}{number of SDs to take from mean if unimodal/one-class}

\item{bimodal_nsd}{number of SDs to take from mean if bimodal/multi-class}
}
\value{
Value to use as a cutoff
}
\description{
Automate Threshold
}
