% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_for_cutoff.R
\name{class_for_cutoff}
\alias{class_for_cutoff}
\title{Class for Cutoff}
\usage{
class_for_cutoff(cell_name, predictions)
}
\arguments{
\item{cell_name}{the name of the cell}

\item{predictions}{the dataframe of predictions}
}
\value{
class to use in automatic threshold function
}
\description{
Class for Cutoff
}
