% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_cells_RNA_no_scale.R
\name{classify_cells_RNA_no_scale}
\alias{classify_cells_RNA_no_scale}
\title{classify_cells_RNA_no_scale}
\usage{
classify_cells_RNA_no_scale(
  imputation,
  model,
  genes_in_model,
  cells_to_use,
  normalized_counts,
  mc.cores = (parallel::detectCores() - 1)
)
}
\arguments{
\item{imputation}{TRUE or FALSE for whether to perform imputation}

\item{model}{Path to RF model to use}

\item{genes_in_model}{Vector of genes to use in the model}

\item{cells_to_use}{Vector of cell names to use from the rna_counts, refers to their column names}

\item{normalized_counts}{library-size normalized counts}

\item{mc.cores}{number of cores}
}
\value{
A matrix with predictions of each cell class in the model
}
\description{
classify_cells_RNA_no_scale
}
