% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_layer.R
\name{classify_layer}
\alias{classify_layer}
\title{classify_layer}
\usage{
classify_layer(
  rna_counts,
  cells_to_use,
  imputation = TRUE,
  genes_in_model,
  model,
  ref_based = F,
  mc.cores = (parallel::detectCores() - 1),
  unimodal_nsd = 3,
  bimodal_nsd = 2,
  layer,
  normalized_counts
)
}
\arguments{
\item{rna_counts}{count matrix}

\item{cells_to_use}{vector indicating which cells to classify in layer}

\item{imputation}{whether to apply MAGIC imputation - recommended}

\item{genes_in_model}{genes used in the model}

\item{model}{model}

\item{ref_based}{whether to use ref based - currently depreceated}

\item{mc.cores}{number of cores}

\item{unimodal_nsd}{number of sd for setting threshold in unimodal distributions}

\item{bimodal_nsd}{number of sd for setting threshold in bimodal distributions}

\item{layer}{which layer is being classified}
}
\value{

}
\description{
classify_layer
}
