% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copykat_no_heatmap.R
\name{copy_kat_no_heatmap}
\alias{copy_kat_no_heatmap}
\title{copy_kat_no_heatmap}
\usage{
copy_kat_no_heatmap(
  rawmat = rawdata,
  summary_matrix,
  id.type = "S",
  cell.line = "no",
  ngene.chr = 5,
  LOW.DR = 0.05,
  UP.DR = 0.1,
  win.size = 25,
  KS.cut = 0.1,
  sam.name = "",
  distance = "euclidean",
  n.cores = (detectCores() - 1)
)
}
\arguments{
\item{rawmat}{matrix of counts}

\item{id.type}{set S if gene names}

\item{cell.line}{set to no for tissues}

\item{ngene.chr}{minimum nchrm, for my tool i use 0 as i want every cell included}

\item{LOW.DR}{minimal population fractions of genes for smoothing.}

\item{UP.DR}{minimal population fractions of genes for segmentation.}

\item{win.size}{minimal window sizes for segmentation.}

\item{KS.cut}{segmentation parameters, input 0 to 1; larger looser criteria.}

\item{sam.name}{sample name.}

\item{distance}{default eucledian}

\item{n.cores}{for parallel computing}

\item{norm.cell.names}{vector of normal cells use the blood and stromal from the classifier}
}
\value{
returns a CNV matrix for use in create_summary_matrix
}
\description{
copy_kat_no_heatmap
}
