% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_auto_threshold.R
\name{get_auto_threshold}
\alias{get_auto_threshold}
\title{get_auto_threshold}
\usage{
get_auto_threshold(
  predictions,
  mc.cores = (parallel::detectCores() - 1),
  unimodal_nsd = 3,
  bimodal_nsd = 2
)
}
\arguments{
\item{predictions}{dataframe of scores and predicted class}

\item{mc.cores}{number of cores for parallel computing}

\item{unimodal_nsd}{number of standard deviations if scores are unimodal}

\item{bimodal_nsd}{number of standard deviations if scores are multimodal}
}
\value{
returns a confidence threshold for each cell type class
}
\description{
get_auto_threshold
}
