% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_scATOMIC_layer.R
\name{get_new_scATOMIC_layer}
\alias{get_new_scATOMIC_layer}
\title{get_new_scATOMIC_layer}
\usage{
get_new_scATOMIC_layer(
  training_data,
  cell_type_metadata,
  output_dir,
  layer_name,
  mc_cores,
  n_cells_replicate = 5000,
  n_trees = 500,
  custom_gene_list = NULL
)
}
\arguments{
\item{training_data}{count matrix, preferably sparse matrix with rows being genes and columns being cells}

\item{cell_type_metadata}{vector containing the cell type corresponding to each cell in training_data}

\item{output_dir}{directory to save random forest and selected genes to}

\item{layer_name}{name of resulting layer}

\item{mc_cores}{number of cores for parallel computing, we suggest using at least 5 cores to speed up process}

\item{n_cells_replicate}{number of cells for balancing each class, we suggest 10000 if <5 cell types, 500 if < 20 cell types and 2500 if > 20}

\item{n_trees}{number of trees in random forest}

\item{custom_gene_list}{vector of features to use, if user wants to provide their own genes.}
}
\value{
list of new rf model and selected features
}
\description{
get_new_scATOMIC_layer
}
\examples{
\dontrun{
#visit github for more information on how to generate training datasets.
breast_cancer_subclassification <- get_new_scATOMIC_layer(training_data = Wu_et_al_breast_count_mat,cell_type_metadata = Wu_et_al_breast_metadata$subtype,
output_dir = "Wu_etal_2021_BRCA_scRNASeq/breast_subclassification_layer/",
mc_cores = 6, layer_name = "breast_subclassification", n_cells_replicate = 10000, n_trees = 500)




}
}
