% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outersect.R
\name{outersect}
\alias{outersect}
\title{outersect}
\usage{
outersect(x, y)
}
\arguments{
\item{x}{first element}

\item{y}{second element}
}
\value{
vector of non overlapping features - opposite of interesect
}
\description{
outersect
}
