% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scATOMIC.R
\name{run_scATOMIC}
\alias{run_scATOMIC}
\title{Run scATOMIC}
\usage{
run_scATOMIC(
  rna_counts,
  imputation = TRUE,
  ref_based = F,
  mc.cores = (parallel::detectCores() - 1),
  unimodal_nsd = 3,
  bimodal_nsd = 2,
  breast_mode = F,
  confidence_cutoff = T,
  fine_grained_T = T
)
}
\arguments{
\item{rna_counts}{count matrix}

\item{imputation}{whether to apply MAGIC imputation - recommended}

\item{ref_based}{whether to use ref based - currently depreceated}

\item{mc.cores}{number of cores}

\item{unimodal_nsd}{number of sd for setting threshold in unimodal distributions}

\item{bimodal_nsd}{number of sd for setting threshold in bimodal distributions}

\item{breast_mode}{run breast subclassfication}

\item{confidence_cutoff}{logical to use confidence cutoffs}

\item{fine_grained_T}{run fine grained T cell subclassification}
}
\value{
list of layers classified
}
\description{
Run scATOMIC
}
\examples{
\dontrun{
#load data
lung_cancer_demo_data <- demo_lung_data
# Quality control filtering
pct_mt <- colSums(lung_cancer_demo_data[grep("^MT-", row.names(lung_cancer_demo_data)),])/colSums(lung_cancer_demo_data) * 100
nFeatureRNA <- apply(lung_cancer_demo_data, 2,function(x){length(which(x != 0))})
lung_cancer_demo_data <- lung_cancer_demo_data[, names(which(pct_mt < 25))]
lung_cancer_demo_data <- lung_cancer_demo_data[, intersect(names(which(nFeatureRNA > 500)), colnames(lung_cancer_demo_data))]
cell_predictions <- run_scATOMIC(lung_cancer_demo_data)
}
}
