% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatomic_tree.R
\name{scATOMICTree}
\alias{scATOMICTree}
\title{scATOMICTree}
\usage{
scATOMICTree(
  predictions_list,
  summary_matrix,
  interactive_mode = T,
  collapsed = T,
  save_results = T,
  save_dir = getwd(),
  project_name = "test_project",
  width = NULL,
  height = NULL,
  fontSize = 10,
  linkLength = 150,
  ncell_size = T
)
}
\arguments{
\item{predictions_list}{list of layered predictions, from run_scATOMIC()}

\item{summary_matrix}{summary matrix of results, from create_summary_matrix()}

\item{interactive_mode}{whether to return collapsibleTree chart, if F flow chart returned}

\item{collapsed}{if interactive mode is T, whether to collapse nodes or open all nodes}

\item{save_results}{logical if you want to save the tree as an html file}

\item{save_dir}{directory to save html file, we recommend working directory}

\item{project_name}{prefix name of project for save file, saved as project_name_annotation_results_tree.html}

\item{width}{width in pixels (optional, defaults to automatic sizing)}

\item{height}{height in pixels (optional, defaults to automatic sizing)}

\item{fontSize}{font size of the label text in pixels}

\item{linkLength}{length of the horizontal links that connect nodes in pixels}

\item{ncell_size}{logical whether to scale size of dots in interactive mode by number of cells (set to false if using relatively homogeneous dataset)}
}
\value{

}
\description{
scATOMICTree
}
\examples{
\dontrun{
#load data
lung_cancer_demo_data <- demo_lung_data
# Quality control filtering
pct_mt <- colSums(lung_cancer_demo_data[grep("^MT-", row.names(lung_cancer_demo_data)),])/colSums(lung_cancer_demo_data) * 100
nFeatureRNA <- apply(lung_cancer_demo_data, 2,function(x){length(which(x != 0))})
lung_cancer_demo_data <- lung_cancer_demo_data[, names(which(pct_mt < 25))]
lung_cancer_demo_data <- lung_cancer_demo_data[, intersect(names(which(nFeatureRNA > 500)), colnames(lung_cancer_demo_data))]
cell_predictions <- run_scATOMIC(lung_cancer_demo_data)
#create results matrix - no CNVs
results_matrix <- create_summary_matrix(prediction_list = cell_predictions, use_CNVs = F, modify_results = T, mc.cores = 1, raw_counts = lung_cancer_demo_data, min_prop = 0.5 )
#visualize on tree
#interactive mode
tree_results_lung <- scATOMICTree(predictions_list = cell_predictions, summary_matrix = results_lung,
interactive_mode = T, collapsed = T, save_results = F,height = 700, width = 1000)
#non interactive mode
tree_results_non_interactive <- scATOMICTree(predictions_list = cell_predictions, summary_matrix = results_lung,
interactive_mode = F, save_results = F)
}
}
