% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_class.R
\name{score_class}
\alias{score_class}
\title{Score Class}
\usage{
score_class(cell_name, predictions, layer)
}
\arguments{
\item{cell_name}{the name of the cell to score}

\item{predictions}{path to the matrix generated from classify cells}

\item{layer}{layer of the model can be layer_1, layer_2_blood, layer_2_cancer, layer_3_TNK, layer_3_MDC, layer_3_BCell, layer_4_CD4_CD8, layer_4_CD8_NK, layer_3_ov_endo, layer_3_breast_lung, layer_3_brain_lung_nbm_skin, layer_4_brain_nbm_skin, layer_5_brain_nbm}
}
\value{
A vector of the predicted class based on scores
}
\description{
Score Class
}
