% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_cells_RNA_no_scale.R
\name{classify_cells_RNA_no_scale}
\alias{classify_cells_RNA_no_scale}
\title{classify_cells_RNA_no_scale}
\usage{
classify_cells_RNA_no_scale(
  rna_counts,
  imputation,
  model,
  genes_in_model,
  cells_to_use,
  ref_based,
  normalized_counts
)
}
\arguments{
\item{rna_counts}{A matrix from the SCT counts slot of a seurat sctransformed object}

\item{imputation}{TRUE or FALSE for whether to perform imputation}

\item{model}{Path to RF model to use}

\item{genes_in_model}{Vector of genes to use in the model}

\item{cells_to_use}{Vector of cell names to use from the rna_counts, refers to their column names}

\item{ref_based}{TRUE or FALSE whether reference based imputation will be used}

\item{normalized_counts}{library-size normalized counts}
}
\value{
A matrix with predictions of each cell class in the model
}
\description{
classify_cells_RNA_no_scale
}
