% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_matrix.R
\name{create_summary_matrix}
\alias{create_summary_matrix}
\title{create_summary_matrix_pan_can_v3}
\usage{
create_summary_matrix(
  raw_counts,
  prediction_list,
  use_CNVs = FALSE,
  modify_results = TRUE,
  mc.cores = 1,
  min_prop = 0.5,
  breast_mode = F,
  fine_grained_T = T,
  confidence_cutoff = T,
  pan_cancer = F,
  cancer_confidence = "default",
  normal_tissue = F,
  low_res_mode = FALSE,
  known_cancer_type = NULL
)
}
\arguments{
\item{raw_counts}{gene by cell count matrix}

\item{prediction_list}{output from run_classifier()}

\item{use_CNVs}{logical whether to use CNVs inference to validate cancer cluster - note takes long time}

\item{modify_results}{logical whether to assume one cancer in the sample and reassign minority cancer calls as Unknown cells}

\item{mc.cores}{number of cores to use in parallel analysis}

\item{min_prop}{minimum \% of major cancer class for modifications}

\item{breast_mode}{run breast subclassfication}

\item{fine_grained_T}{run T cell subclassfication}

\item{confidence_cutoff}{logical to use confidence cutoffs}

\item{pan_cancer}{logical to use pan cancer features or cancer specific features}

\item{cancer_confidence}{assesing whether a cancer prediction should be interpreted as confident. by default we use layer specific cutoff derived from our paper that optimally split true positives and true negatives. Set to a numeric from 0-1 to set a hard cutoff for all layers.}

\item{normal_tissue}{Set to true if you know the sample is only normal cells, will convert all cancer annotations to normal tissue cell}

\item{low_res_mode}{Set to true to convert all high res subtypes into low res cell types}

\item{known_cancer_type}{converts all cancer cell type annotations to provided string, recommended if one knows with 100\% certainty what the cancer type is.}
}
\value{
dataframe of cell names and final prediction
}
\description{
create_summary_matrix_pan_can_v3
}
\examples{
\dontrun{
#load data
lung_cancer_demo_data <- demo_lung_data
# Quality control filtering
pct_mt <- colSums(lung_cancer_demo_data[grep("^MT-", row.names(lung_cancer_demo_data)),])/colSums(lung_cancer_demo_data) * 100
nFeatureRNA <- apply(lung_cancer_demo_data, 2,function(x){length(which(x != 0))})
lung_cancer_demo_data <- lung_cancer_demo_data[, names(which(pct_mt < 25))]
lung_cancer_demo_data <- lung_cancer_demo_data[, intersect(names(which(nFeatureRNA > 500)), colnames(lung_cancer_demo_data))]
cell_predictions <- run_scATOMIC(lung_cancer_demo_data)
#create results matrix - no CNVs
summary_masterrix <- create_summary_matrix(prediction_list = cell_predictions, use_CNVs = F, modify_results = T, mc.cores = 1, raw_counts = lung_cancer_demo_data, min_prop = 0.5 )
#create results matrix with CNV corrections
results_lung_CNV <- create_summary_matrix(prediction_list = cell_predictions, use_CNVs = T, modify_results = T, mc.cores = 6, raw_counts = lung_cancer_demo_data, min_prop = 0.5 )
}
}
